Fix imotioncube types to resemble the CoE manual

+++ march_hardware/include/march_hardware/PDOmap.h
@@ -68,7 +68,7 @@ public:
    */
   void addObject(IMCObjectName object_name);
 
-  std::unordered_map<IMCObjectName, int> map(int slave_index, DataDirection direction);
+  std::unordered_map<IMCObjectName, uint8_t> map(int slave_index, DataDirection direction);
 
   static std::unordered_map<IMCObjectName, IMCObject> all_objects;
 
@@ -79,7 +79,7 @@ private:
 
   /** Configures the PDO in the IMC using the given base register address and sync manager address.
    * @return map of the IMC PDO object name in combination with the byte-offset in the PDO register */
-  std::unordered_map<IMCObjectName, int> configurePDO(int slave_index, int base_register, int base_sync_manager);
+  std::unordered_map<IMCObjectName, uint8_t> configurePDO(int slave_index, int base_register, int base_sync_manager);
 
   std::unordered_map<IMCObjectName, IMCObject> PDO_objects;
   int total_used_bits = 0;
